Add CODEOWNERS parsing functionality to gh
