+++ .github/CONTRIBUTING
@@ -1,0 +1,35 @@
+# Contributing
+
+So you want to code? Awesome. The following is the contributing guide for Elhub development inner-source code. Please read these guidelines if you haven't
+done so before. By following these guidelines, you will save the owning team both time and effort, and they will be better able to help you in addressing
+your issues, assess changes, and helping you finalize any contributions to the project.
+
+You can contribute in many ways to the project:
+
+* Submit bug reports or feature requests
+* Improve the documentation, either the README here or the documentation on confluence (see README for link)
+* Hack on the project itself by fixing bugs you've found or adding new features
+
+If you have more general questions about this project, use the Team Dev Teams channel.
+
+## How to Report a bug or suggest a feature/enchancement
+
+If you are an Elhub employee, follow the [guidelines on Confluence](https://confluence.elhub.cloud/display/DEV/How+to+Submit+a+Bug+or+Feature+Request).
+
+If you are accessing this repository on Github, submit an issue using the project's issue board on Github.
+
+## How to Fix a bug or Add a feature/enhancement
+
+If you are an Elhub employee, follow the [guidelines on Confluence](https://confluence.elhub.cloud/pages/viewpage.action?pageId=641763474).
+
+If you are looking at this repository on Github, feel free to submit a PR. Is this new to you? You can learn how to do it from this *free* set of tutorials:
+[How to Contribute to an Open Source Project on GitHub](https://egghead.io/series/how-to-contribute-to-an-open-source-project-on-github).
+
+All code that is submitted into the project is code-reviewed by the responsible team (listed in CODEOWNERS) before being pushed into the repository.
+
+Expectations:
+
+* Create an issue for any major change or enhancement before writing any code. Creating an issue allows the change to be discussed openly by the team.
+* Ensure that code submitted follow our code guidelines (running the linters will ensure this).
+* Keep commits as small as possible, preferably one change per commit.
+* Be considerate and respectful to others.
